// src/models/search.ts
//import  moment  from 'moment';

export class Search {
    
    originLocationName: string='Jakarta - Cengkareng';
    // Origin Location 3 Code
    originLocationAlpha3Code: string='CGK';
    lat: number =0;
    lng: number=0;
    querysearch:string='';
	dstart: any=new Date();
	dend: any=new Date();
}